Add support for CC2530, and update PyUSB usage
This commit adds support for the CC2530 USB dongle, which has
different USB vendor/product IDs, and also uses endpoint 0x82
rather than 0x83 for data transfer.

If a CC2531 USB device is not found, then an attempt to detect
a CC2530 device is made instead - so either dongle will be
detected.